Rename module from src to biomassrecovery
