+++ biomassrecovery/data/gedi_database_loader.py
@@ -10,10 +10,10 @@ import sqlalchemy as db
 from sqlalchemy.engine.base import Engine
 from sqlalchemy.sql.schema import Table
 import warnings
-from src.data.gedi_granule import GediGranule
-from src.utils import logging_util
-from src.constants import WGS84
-from src.environment import DB_CONFIG
+from biomassrecovery.data.gedi_granule import GediGranule
+from biomassrecovery.utils import logging_util
+from biomassrecovery.constants import WGS84
+from biomassrecovery.environment import DB_CONFIG
 
 logger = logging_util.get_logger("data_logger", INFO)
