+++ biomassrecovery/data/jrc_download.py
@@ -10,8 +10,8 @@ import subprocess
 import psutil
 import requests
 
-from src.environment import DATA_PATH
-from src.utils.logging_util import get_logger
+from biomassrecovery.environment import DATA_PATH
+from biomassrecovery.utils.logging_util import get_logger
 
 logger = get_logger(__name__)
