+++ biomassrecovery/data/jrc_loading.py
@@ -8,9 +8,9 @@ import rioxarray as rxr
 import xarray as xr
 from tqdm.autonotebook import tqdm
 
-from src.environment import JRC_PATH
-from src.utils.logging_util import get_logger
-from src.utils.os import list_content
+from biomassrecovery.environment import JRC_PATH
+from biomassrecovery.utils.logging_util import get_logger
+from biomassrecovery.utils.os import list_content
 
 logger = get_logger(__file__)
