+++ biomassrecovery/utils/logging_util.py
@@ -4,7 +4,7 @@ import os
 import sys
 from logging.handlers import TimedRotatingFileHandler
 
-from src.environment import DEFAULT_FORMATTER, DEFAULT_LOG_FILE, DEFAULT_LOG_LEVEL
+from biomassrecovery.environment import DEFAULT_FORMATTER, DEFAULT_LOG_FILE, DEFAULT_LOG_LEVEL
 
 
 def get_console_handler(
