+++ setup.py
@@ -1,8 +1,8 @@
 from setuptools import find_packages, setup
 
 setup(
-    name="src",
-    version="0.0.1",
+    name="biomassrecovery",
+    version="0.0.2",
     author="Simon V. Mathis (AI4ER CDT, University of Cambridge)",
     author_email="author@example.com",
     description="A short description of the project.",
